Train, Evaluate and Compare the models
Created a function that takes models as dictionary
and training and testing data, then fits the data
into each model and tests the models and stores the
scores of each model as dictionary and finally
returns the score.

Compared the scores of each model using bar graph,
found that the K-Nearest Neighbour Classifier
model performed the worst.